requirements.txt
pyyaml>=6.0
feedparser>=6.0
httpx>=0.27
pydantic>=2.0
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
apscheduler>=3.10
Jinja2>=3.1
python-jose[cryptography]>=3.3
ldap3>=2.9
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
respx>=0.21
